requirements/dev.txt
-r base.txt
dj-database-url==2.1.0
gunicorn==23.0.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3